Update Student.java
Alright hopefully this is the last time I have to add it. This is the actual file i meant to upload. Didn't mean to stress you out with all my errors. Also if i messed up with the syncing I just want to post this so its not completely deleted. Is that the correct way to access the gethours function from each list element?