Add files via upload
Code works for grades so far for grades. Just have to add an "addCourse" function into course class and "addGrade" function in grades class so that it can be used in the code. The only issue with course controller is the time variable. Since were working with interval variable, I'm not sure how to convert datepicker value to interval value. Maybe we should set it up as two strings or ints, using start and end time?